move choose_characters() to functions.py
